changes in editor - using size_t instead of WORD for text length

+++ Editor/OldEditor.h
@@ -4,7 +4,6 @@
 #include "../cppfand/base.h"
 #include "../cppfand/constants.h"
 #include "../cppfand/pstring.h"
-#include "../cppfand/Rdb.h"
 
 class Instr;
 struct EdExitD;
@@ -16,11 +15,11 @@ struct MsgStr
 };
 
 bool EditText(char pMode, char pTxtType, pstring pName, pstring pErrMsg,
-	char* pTxtPtr, WORD pMaxLen, WORD& pLen, WORD& pInd, longint& pScr,
+	char* pTxtPtr, WORD pMaxLen, size_t& pLen, WORD& pInd, longint& pScr,
 	pstring pBreaks, EdExitD* pExD, bool& pSrch, bool& pUpdat,
 	WORD pLastNr, WORD pCtrlLastNr, MsgStr* pMsgS); // r169
 void SimpleEditText(char pMode, pstring pErrMsg, pstring pName, char* TxtPtr,
-	WORD MaxLen, WORD& Len, WORD& Ind, bool& Updat); // r202
+	WORD MaxLen, size_t& Len, WORD& Ind, bool& Updat); // r202
 WORD FindTextE(const pstring& PstrScreenStr, pstring Popt, char* PTxtPtr, WORD PLen); // r209
 void InitTxtEditor();
 void EditTxtFile(longint* LP, char Mode, pstring& ErrMsg, EdExitD* ExD, longint TxtPos,
